Remove net and procutils config from init
Removes the configuration of the processutils and linux_net modules
during os_vif.initialize(**config), since these modules will not
actually be in os_vif, but will rather be in the VIF plugin repositories
themselves (see vif_plug_ovs for an example).

Add docstring to os_vif.initialize(**config) that explains the
configuration options that are passed to the VIF plugin classes.