Add os_vif.objects.InstanceInfo model
One thing that occurred to me when going to add the InstanceConfig
object from Dan's original pull request [1] was that it would be easier
to have only a single argument to the library's plug() and unplug()
method -- the os_vif.objects.VIF object only. The reason is because the
plugin describe() interface now returns a PluginInfo object that
describes the name of the plugin and the min and max object version that
the plugin understands. This object version is the VIF object version,
however the original design had a nova.objects.Instance object being
passed into the plug() and unplug() method. There was nothing in the
plugin describe() contract that indicated what version of the
nova.objects.Instance object the plugin understood, so I thought it
would be best to make a slimmed-down InstanceInfo object be an
ObjectField on the main VIF object and in that manner ensure that the
VIF object version could be the only version information that the
PluginInfo object would need to maintain.

[1]
https://github.com/berrange/os_vif/commit/18585c42b256159c85eeb1f8481a834385892298